Restructured phot_db tables as agreed to allow for multiple filter reference images and traceability.  Resolved issues with ingestion of integer values

+++ pyDANDIA/tests/test_phot_db.py
@@ -43,10 +43,12 @@ def test_get_connection():
     test_conn = sqlite3.connect(db_file_path)
         
     assert type(conn) == type(test_conn)
+    conn.close()
 
 def test_feed_to_table_many():
 
-    os.remove(db_file_path)
+    if os.path.isfile(db_file_path):
+        os.remove(db_file_path)
         
     conn = phot_db.get_connection(dsn=db_file_path)
     
@@ -58,33 +60,39 @@ def test_feed_to_table_many():
     t = phot_db.query_to_astropy_table(conn, query, args=())
     
     assert len(t) == len(tuples)
+    conn.close()
 
 def test_ingest_astropy_table():
     
-    os.remove(db_file_path)
+    if os.path.isfile(db_file_path):
+        os.remove(db_file_path)
         
     conn = phot_db.get_connection(dsn=db_file_path)
     
     n_entries = 2000
     
     data = [ table.Column(name='ra', data=np.linspace(269.0,272.0,n_entries)),
-              table.Column(name='dec', data=np.linspace(-27.5, -28.9, n_entries)) ]
-              
+              table.Column(name='dec', data=np.linspace(-27.5, -28.9, n_entries)),
+              table.Column(name='type', data=['dwarf']*n_entries) ]
+    
     stars = table.Table(data=data)
 
-    phot_db.ingest_astropy_table(conn, 'Stars', stars)
+    phot_db.ingest_astropy_table(conn, 'stars', stars)
     
-    query = 'SELECT star_id, ra, dec FROM stars'
+    query = 'SELECT star_id, ra, dec,type FROM stars'
     t = phot_db.query_to_astropy_table(conn, query, args=())
-    
+
     assert len(t) == n_entries
+    conn.close()
     
 def test_ingest_reference_in_db():
     """Function to test the ingestion of a RefereinceImage into the Phot DB"""
     
-    os.remove(db_file_path)
+    if os.path.isfile(db_file_path):
+        os.remove(db_file_path)
         
     setup = pipeline_setup.pipeline_setup({'red_dir': TEST_DIR})
+    setup.phot_db_path = db_file_path
     
     conn = phot_db.get_connection(dsn=setup.phot_db_path)
     
@@ -103,19 +111,27 @@ def test_ingest_reference_in_db():
     i = np.where(ref_image_name == meta.headers_summary[1]['IMAGES'].data)[0][0]
     
     ref_header = meta.headers_summary[1][i]
-
+    
+    field_id = 'ROME-TEST-01'
+    version = 'Code_test'
+    
     phot_db.ingest_reference_in_db(conn, setup, ref_header, 
                                                 ref_image_dir, 
-                                                ref_image_name)
+                                                ref_image_name,
+                                                field_id, version)
 
-    query = 'SELECT refimg_name FROM reference_images'
+    query = 'SELECT refimg_name,field_id,current_best FROM reference_images WHERE refimg_name="'+ref_image_name+'"'
     t = phot_db.query_to_astropy_table(conn, query, args=())
     
     assert t[0]['refimg_name'] == ref_image_name
+    assert t[0]['current_best'] == 1
 
+    conn.close()
+    
 def test_box_search_on_position():
     
-    os.remove(db_file_path)
+    if os.path.isfile(db_file_path):
+        os.remove(db_file_path)
         
     conn = phot_db.get_connection(dsn=db_file_path)
     
@@ -131,12 +147,48 @@ def test_box_search_on_position():
     results = phot_db.box_search_on_position(conn, ra_centre, dec_centre, dra, ddec)
     
     assert len(results) == len(tuples)
+
+    conn.close()
+    
+def update_table_entry():
+    
+    if os.path.isfile(db_file_path):
+        os.remove(db_file_path)
+        
+    conn = phot_db.get_connection(dsn=db_file_path)
+    
+    (names, tuples) = generate_test_catalog()
+    
+    phot_db.feed_to_table_many(conn, 'Stars', names, tuples)
+    
+    query = 'SELECT star_id,ra,dec FROM stars'
+    t = phot_db.query_to_astropy_table(conn, query, args=())
+    print(t)
+    
+    table_name = 'Stars'
+    key_name = 'ra'
+    search_key = 'star_id'
+    entry_id = 1
+    value = 18.000
+    
+    phot_db.update_table_entry(conn,table_name,key_name,search_key,entry_id,value)
+    
+    query = 'SELECT star_id,ra,dec FROM stars'
+    t = phot_db.query_to_astropy_table(conn, query, args=())
+    print(t)
+    
+    idx = np.where(t['star_id'] == entry_id)
+
+    assert t['ra'][idx].data == value
+    
+    conn.close()
     
 if __name__ == '__main__':
     
     #test_get_connection()
     #test_feed_to_table_many()
-    test_ingest_astropy_table()
-    #test_ingest_reference_in_db()
+    #test_ingest_astropy_table()
+    test_ingest_reference_in_db()
     #test_box_search_on_position()
+    #update_table_entry()
